+++ HPXMLtoOpenStudio/tests/test_defaults.rb
@@ -1693,6 +1693,8 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.hvac_distributions[0].number_of_return_registers = 2
     hpxml.hvac_distributions[0].ducts[0].duct_surface_area_multiplier = 0.5
     hpxml.hvac_distributions[0].ducts[1].duct_surface_area_multiplier = 1.5
+    hpxml.hvac_distributions[0].ducts[0].duct_buried_insulation_level = HPXML::DuctBuriedInsulationPartial
+    hpxml.hvac_distributions[0].ducts[1].duct_buried_insulation_level = HPXML::DuctBuriedInsulationDeep
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
     expected_supply_locations = ['attic - unvented']
@@ -1703,9 +1705,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [1.0]
     expected_supply_area_mults = [0.5]
     expected_return_area_mults = [1.5]
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationPartial]
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationDeep]
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ conditioned basement
     hpxml.hvac_distributions[0].number_of_return_registers = nil
@@ -1713,6 +1718,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       duct.duct_location = nil
       duct.duct_surface_area = nil
       duct.duct_surface_area_multiplier = nil
+      duct.duct_buried_insulation_level = nil
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
@@ -1724,9 +1730,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [1.0]
     expected_supply_area_mults = [1.0]
     expected_return_area_mults = [1.0]
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone]
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone]
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ multiple foundations
     hpxml = _create_hpxml('base-foundation-multiple.xml')
@@ -1736,6 +1745,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       duct.duct_location = nil
       duct.duct_surface_area = nil
       duct.duct_surface_area_multiplier = nil
+      duct.duct_buried_insulation_level = nil
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
@@ -1747,9 +1757,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [1.0]
     expected_supply_area_mults = [1.0]
     expected_return_area_mults = [1.0]
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone]
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone]
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ foundation exposed to ambient
     hpxml = _create_hpxml('base-foundation-ambient.xml')
@@ -1759,6 +1772,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       duct.duct_location = nil
       duct.duct_surface_area = nil
       duct.duct_surface_area_multiplier = nil
+      duct.duct_buried_insulation_level = nil
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
@@ -1770,9 +1784,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [1.0]
     expected_supply_area_mults = [1.0]
     expected_return_area_mults = [1.0]
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone]
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone]
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ building/unit adjacent to other housing unit
     hpxml = _create_hpxml('base-bldgtype-multifamily-adjacent-to-other-housing-unit.xml')
@@ -1782,6 +1799,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       duct.duct_location = nil
       duct.duct_surface_area = nil
       duct.duct_surface_area_multiplier = nil
+      duct.duct_buried_insulation_level = nil
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
@@ -1793,9 +1811,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [1.0]
     expected_supply_area_mults = [1.0]
     expected_return_area_mults = [1.0]
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone]
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone]
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ 2-story building
     hpxml = _create_hpxml('base-enclosure-2stories.xml')
@@ -1805,6 +1826,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       duct.duct_location = nil
       duct.duct_surface_area = nil
       duct.duct_surface_area_multiplier = nil
+      duct.duct_buried_insulation_level = nil
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
@@ -1816,9 +1838,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [0.375, 0.375, 0.125, 0.125]
     expected_supply_area_mults = [1.0, 1.0, 1.0, 1.0]
     expected_return_area_mults = [1.0, 1.0, 1.0, 1.0]
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone] * 4
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone] * 4
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ 1-story building & multiple HVAC systems
     hpxml = _create_hpxml('base-hvac-multiple.xml')
@@ -1831,6 +1856,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
         duct.duct_location = nil
         duct.duct_surface_area = nil
         duct.duct_surface_area_multiplier = nil
+        duct.duct_buried_insulation_level = nil
       end
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
@@ -1843,9 +1869,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [0.5, 0.5] * hpxml_default.hvac_distributions.size
     expected_supply_area_mults = [1.0, 1.0] * hpxml_default.hvac_distributions.size
     expected_return_area_mults = [1.0, 1.0] * hpxml_default.hvac_distributions.size
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone] * 2 * hpxml_default.hvac_distributions.size
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone] * 2 * hpxml_default.hvac_distributions.size
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ 2-story building & multiple HVAC systems
     hpxml = _create_hpxml('base-hvac-multiple.xml')
@@ -1859,6 +1888,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
         duct.duct_location = nil
         duct.duct_surface_area = nil
         duct.duct_surface_area_multiplier = nil
+        duct.duct_buried_insulation_level = nil
       end
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
@@ -1871,9 +1901,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [0.375, 0.375, 0.125, 0.125] * hpxml_default.hvac_distributions.size
     expected_supply_area_mults = [1.0, 1.0, 1.0, 1.0] * hpxml_default.hvac_distributions.size
     expected_return_area_mults = [1.0, 1.0, 1.0, 1.0] * hpxml_default.hvac_distributions.size
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone] * 4 * hpxml_default.hvac_distributions.size
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone] * 4 * hpxml_default.hvac_distributions.size
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
 
     # Test defaults w/ 2-story building & multiple HVAC systems & duct area fractions
     hpxml = _create_hpxml('base-hvac-multiple.xml')
@@ -1892,6 +1925,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       hvac_distribution.ducts.each do |duct|
         duct.duct_surface_area = nil
         duct.duct_surface_area_multiplier = nil
+        duct.duct_buried_insulation_level = nil
       end
     end
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
@@ -1904,9 +1938,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     expected_return_fracs = [0.5, 0.5] * hpxml_default.hvac_distributions.size
     expected_supply_area_mults = [1.0, 1.0] * hpxml_default.hvac_distributions.size
     expected_return_area_mults = [1.0, 1.0] * hpxml_default.hvac_distributions.size
+    expected_supply_buried_levels = [HPXML::DuctBuriedInsulationNone] * 2 * hpxml_default.hvac_distributions.size
+    expected_return_buried_levels = [HPXML::DuctBuriedInsulationNone] * 2 * hpxml_default.hvac_distributions.size
     expected_n_return_registers = hpxml_default.building_construction.number_of_conditioned_floors
     _test_default_duct_values(hpxml_default, expected_supply_locations, expected_return_locations, expected_supply_areas, expected_return_areas,
-                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults)
+                              expected_supply_fracs, expected_return_fracs, expected_n_return_registers, expected_supply_area_mults, expected_return_area_mults,
+                              expected_supply_buried_levels, expected_return_buried_levels)
   end
 
   def test_mech_ventilation_fans
@@ -4036,7 +4073,8 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
   end
 
   def _test_default_duct_values(hpxml, supply_locations, return_locations, supply_areas, return_areas,
-                                supply_fracs, return_fracs, n_return_registers, supply_area_mults, return_area_mults)
+                                supply_fracs, return_fracs, n_return_registers, supply_area_mults, return_area_mults,
+                                supply_buried_levels, return_buried_levels)
     supply_duct_idx = 0
     return_duct_idx = 0
     hpxml.hvac_distributions.each do |hvac_distribution|
@@ -4049,12 +4087,14 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
           assert_in_epsilon(supply_areas[supply_duct_idx], duct.duct_surface_area, 0.01)
           assert_in_epsilon(supply_fracs[supply_duct_idx], duct.duct_fraction_area, 0.01)
           assert_in_epsilon(supply_area_mults[supply_duct_idx], duct.duct_surface_area_multiplier, 0.01)
+          assert_equal(supply_buried_levels[supply_duct_idx], duct.duct_buried_insulation_level)
           supply_duct_idx += 1
         elsif duct.duct_type == HPXML::DuctTypeReturn
           assert_equal(return_locations[return_duct_idx], duct.duct_location)
           assert_in_epsilon(return_areas[return_duct_idx], duct.duct_surface_area, 0.01)
           assert_in_epsilon(return_fracs[return_duct_idx], duct.duct_fraction_area, 0.01)
           assert_in_epsilon(return_area_mults[return_duct_idx], duct.duct_surface_area_multiplier, 0.01)
+          assert_equal(return_buried_levels[return_duct_idx], duct.duct_buried_insulation_level)
           return_duct_idx += 1
         end
       end